Share more processes code between macos and freebsd
